+++ BasicNetworking/Models/ComicCharacter.swift
