+++ BasicNetworking/Models/MarvelResponse.swift
