+++ BasicNetworking/NetworkError.swift
@@ -1,0 +1,12 @@
+//
+//  NetworkError.swift
+//  BasicNetworking
+//
+//  Created by Fernando on 19.04.19.
+//
+
+import Foundation
+
+enum NetworkError: Error {
+    case invalidURL
+}
